+++ src/AssemblyInfo.cs
@@ -1,0 +1,9 @@
+using System.Reflection;
+
+[assembly: AssemblyTitle("Windows-Wartung")]
+[assembly: AssemblyProduct("Windows-Wartung")]
+[assembly: AssemblyDescription("Wartungs- und Reparatur-Toolbox für Windows")]
+[assembly: AssemblyCompany("Jonas")]
+[assembly: AssemblyCopyright("© 2026 Jonas")]
+[assembly: AssemblyVersion("3.0.0.0")]
+[assembly: AssemblyFileVersion("3.0.0.0")]